# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Case API",
    "version": "1.0.0",
    "description": "HM Land Registry Digital Street Proof of Concept",
    "contact": {
      "name": "HM Land Registry"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "summary": "Retrieve a list of Users",
        "operationId": "get_users",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "name": "email_address",
            "in": "query",
            "description": "The email address to filter by",
            "schema": {
              "type": "string"
            },
            "example": "john.doe@example.com"
          }
        ],
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new User",
        "operationId": "create_user",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "description": "New User data to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Newly created User",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "links": {
              "GetUserByIdentity": {
                "$ref": "#/components/links/GetUserByIdentity"
              },
              "UpdateUserByIdentity": {
                "$ref": "#/components/links/UpdateUserByIdentity"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/{identity}": {
      "get": {
        "summary": "Retrieve a specific User",
        "operationId": "get_user",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/identity"
          }
        ],
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "links": {
              "UpdateUserByIdentity": {
                "$ref": "#/components/links/UpdateUserByIdentity"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a specific User",
        "operationId": "update_user",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/identity"
          }
        ],
        "requestBody": {
          "description": "User data to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Expected response to a valid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "links": {
              "GetUserByIdentity": {
                "$ref": "#/components/links/GetUserByIdentity"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "User": {
        "required": [
          "identity",
          "first_name",
          "last_name",
          "email_address",
          "phone_number",
          "address"
        ],
        "properties": {
          "identity": {
            "type": "integer",
            "format": "int32"
          },
          "first_name": {
            "type": "string",
            "example": "Natasha"
          },
          "last_name": {
            "type": "string",
            "example": "Powell"
          },
          "email_address": {
            "type": "string",
            "format": "email",
            "example": "n.powell@example.com"
          },
          "phone_number": {
            "type": "string",
            "example": "07700900354"
          },
          "address": {
            "$ref": "#/components/schemas/UserAddress"
          }
        }
      },
...
      "UserAddress": {
        "required": [
          "house_name_number",
          "street",
          "town_city",
          "county",
          "country",
          "postcode"
        ],
        "properties": {
          "house_name_number": {
            "type": "string",
            "example": 143
          },
          "street": {
            "type": "string",
            "example": "Cowgate"
          },
          "town_city": {
            "type": "string",
            "example": "Edinburgh"
          },
          "county": {
            "type": "string",
            "example": "Midlothian"
          },
          "country": {
            "type": "string",
            "example": "Scotland"
          },
          "postcode": {
            "type": "string",
            "pattern": "^(GIR ?0AA|[A-PR-UWYZ]([0-9]{1,2}|([A-HK-Y][0-9]([0-9ABEHMNPRV-Y])?)|[0-9][A-HJKPS-UW]) ?[0-9][ABD-HJLNP-UW-Z]{2})$",
            "example": "EH1 1JS"
          }
        }
      },
...
      "Error": {
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    },
    "parameters": {
...
      "identity": {
        "name": "identity",
        "in": "path",
        "required": true,
        "description": "The user identity of the user to process",
        "schema": {
          "type": "integer"
        }
      }
    },
    "links": {
...
      "GetUserByIdentity": {
        "operationId": "get_user",
        "parameters": {
          "identity": "$response.body#/identity"
        }
      },
      "UpdateUserByIdentity": {
        "operationId": "update_user",
        "parameters": {
          "identity": "$response.body#/identity"
        }
      }
    }
  }
}